Create a mixin module from renderer class
Move the method definitions in the Renderer class into a mixin that can
be included into a subclass of `Redcardpet::Render::HTML`.

This is useful when trying to customise an existing extension to
Redcarpet, for instance markdown-rails.
